requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
idna==2.7
numpy==1.15.4
pycparser==2.19
pycrypto==2.6.1
pyOpenSSL==18.0.0
PySocks==1.6.8
QtAwesome==0.5.3
QtPy==1.5.2
requests==2.20.1
six==1.12.0
urllib3==1.23
wincertstore==0.2
